Prevent some overflows in WndProc

+++ EDDiscovery/Forms/TabControlForm.cs
@@ -232,7 +232,7 @@ namespace EDDiscovery.Forms
         protected override void WndProc(ref Message m)
         {
             // Compatibility movement for Mono
-            if (m.Msg == WM.LBUTTONDOWN && (int)m.WParam == 1 && !windowsborder)
+            if (m.Msg == WM.LBUTTONDOWN && m.WParam == (IntPtr)1 && !windowsborder)
             {
                 int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                 int y = unchecked((short)((uint)m.LParam >> 16));
@@ -242,7 +242,7 @@ namespace EDDiscovery.Forms
                 m.Result = IntPtr.Zero;
                 this.Capture = true;
             }
-            else if (m.Msg == WM.MOUSEMOVE && (int)m.WParam == 1 && _window_dragging)
+            else if (m.Msg == WM.MOUSEMOVE && m.WParam == (IntPtr)1 && _window_dragging)
             {
                 int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                 int y = unchecked((short)((uint)m.LParam >> 16));
@@ -265,7 +265,7 @@ namespace EDDiscovery.Forms
             {
                 base.WndProc(ref m);
 
-                if ((int)m.Result == HT.CLIENT)
+                if (m.Result == (IntPtr)HT.CLIENT)
                 {
                     int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                     int y = unchecked((short)((uint)m.LParam >> 16));
